refactor(ui): refactor log stream / viewer sse
