+++ services/dashboard-ui/client/components/runners/job-details/RunnerJobLogs/RunnerJobLogs.stories.tsx
@@ -5,5 +5,5 @@ export default {
 import { RunnerJobLogs } from './RunnerJobLogs'
 
 export const NoLogStream = () => (
-  <RunnerJobLogs jobStatus="queued" />
+  <RunnerJobLogs />
 )
