+++ services/dashboard-ui/client/components/runners/job-details/RunnerJobLogs/RunnerJobLogsContainer.tsx
@@ -3,5 +3,5 @@ import { RunnerJobLogs } from './RunnerJobLogs'
 
 export const RunnerJobLogsContainer = () => {
   const { job } = useRunnerJob()
-  return <RunnerJobLogs logStreamId={job.log_stream_id} jobStatus={job.status} />
+  return <RunnerJobLogs logStreamId={job.log_stream_id} />
 }
