+++ services/dashboard-ui/client/components/workflows/step-details/deploy-details/DeployStepDetails/DeployStepDetails.tsx
@@ -10,7 +10,6 @@ import { SSELogs } from '@/components/log-stream/SSELogs'
 import { TraceView } from '@/components/spans/TraceView'
 import { LogStreamProvider } from '@/providers/log-stream-provider'
 import { LogViewerProvider } from '@/providers/log-viewer-provider'
-import { UnifiedLogsProvider } from '@/providers/unified-logs-provider'
 import type { TDeploy, TWorkflowStep } from '@/types'
 import { DeployApply, DeployLogsSkeleton } from '../DeployApply'
 
@@ -84,27 +83,25 @@ export const DeployStepDetails = ({
       </div>
       {step?.execution_type === 'approval' ? (
         deploy?.log_stream ? (
-          <LogStreamProvider shouldPoll logStreamId={deploy.log_stream.id}>
-            <UnifiedLogsProvider>
-              <LogViewerProvider>
-                <Tabs
-                  tabs={{
-                    plan: (
-                      <div className="mt-4">
-                        <Plan step={step} />
-                      </div>
-                    ),
-                    logs: <SSELogs />,
-                    trace: (
-                      <TraceView
-                        logStreamId={deploy.log_stream.id}
-                        shouldPoll={deploy.log_stream.open}
-                      />
-                    ),
-                  }}
-                />
-              </LogViewerProvider>
-            </UnifiedLogsProvider>
+          <LogStreamProvider logStreamId={deploy.log_stream.id}>
+            <LogViewerProvider>
+              <Tabs
+                tabs={{
+                  plan: (
+                    <div className="mt-4">
+                      <Plan step={step} />
+                    </div>
+                  ),
+                  logs: <SSELogs />,
+                  trace: (
+                    <TraceView
+                      logStreamId={deploy.log_stream.id}
+                      shouldPoll={deploy.log_stream.open}
+                    />
+                  ),
+                }}
+              />
+            </LogViewerProvider>
           </LogStreamProvider>
         ) : (
           <Tabs
