+++ services/dashboard-ui/client/hooks/use-logs.ts
@@ -1,13 +1,13 @@
 import { useContext } from 'react'
-import { UnifiedLogsContext } from '@/providers/unified-logs-provider'
+import { LogStreamContext, type LogStreamContextValue } from '@/providers/log-stream-provider'
 import { LogViewerContext } from '@/providers/log-viewer-provider'
 import type { TLogFiltersProps } from '@/hooks/use-log-filters'
 import type { TOTELLog, TAPIError } from '@/types'
 
-export const useUnifiedLogData = () => {
-  const context = useContext(UnifiedLogsContext)
+export const useLogStreamData = (): LogStreamContextValue => {
+  const context = useContext(LogStreamContext)
   if (context === undefined) {
-    throw new Error('useUnifiedLogData must be used within a UnifiedLogsProvider')
+    throw new Error('useLogStreamData must be used within a LogStreamProvider')
   }
   return context
 }
@@ -20,34 +20,7 @@ export const useLogViewer = () => {
   return context
 }
 
-export const useUnifiedLogsComplete = () => {
-  const logData = useUnifiedLogData()
-  const viewer = useLogViewer()
-
-  return {
-    logs: logData.logs,
-    isLoading: logData.isLoading,
-    error: logData.error,
-    connectionState: logData.connectionState,
-    loadMore: logData.loadMore,
-    hasMore: logData.hasMore,
-    isStreamOpen: logData.isStreamOpen,
-    activeLog: viewer.activeLog,
-    filteredLogs: viewer.filteredLogs,
-    filters: viewer.filters,
-    handleActiveLog: viewer.handleActiveLog,
-  }
-}
-
-export type UnifiedLogsContextValue = {
-  logs: TOTELLog[]
-  isLoading: boolean
-  error: TAPIError | null
-  connectionState: 'disconnected' | 'connecting' | 'connected' | 'reconnecting'
-  loadMore: () => void
-  hasMore: boolean
-  isStreamOpen: boolean
-}
+export type { LogStreamContextValue }
 
 export type LogViewerContextValue = {
   activeLog?: TOTELLog
