+++ services/dashboard-ui/client/views/install/deploy-tabs/DeployLogsTab.tsx
@@ -1,7 +1,6 @@
 import { SSELogs } from '@/components/log-stream/SSELogs'
 import { LogStreamProvider } from '@/providers/log-stream-provider'
 import { LogViewerProvider } from '@/providers/log-viewer-provider'
-import { UnifiedLogsProvider } from '@/providers/unified-logs-provider'
 import { useDeploy } from '@/hooks/use-deploy'
 
 export const DeployLogsTab = () => {
@@ -9,12 +8,10 @@ export const DeployLogsTab = () => {
   const logStream = deploy?.log_stream
 
   return (
-    <LogStreamProvider logStreamId={logStream?.id} shouldPoll={logStream?.open}>
-      <UnifiedLogsProvider>
-        <LogViewerProvider>
-          <SSELogs filterClassName="top-0" />
-        </LogViewerProvider>
-      </UnifiedLogsProvider>
+    <LogStreamProvider logStreamId={logStream?.id}>
+      <LogViewerProvider>
+        <SSELogs filterClassName="top-0" />
+      </LogViewerProvider>
     </LogStreamProvider>
   )
 }
